% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_all.R
\name{download_data_raw_ui}
\alias{download_data_raw_ui}
\title{Shiny module ui for downloading data_all}
\usage{
download_data_raw_ui(id = "data_raw")
}
\description{
Shiny module ui for downloading data_all
}
