% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data_server}
\alias{download_data_server}
\title{Shiny module server for downloading data}
\usage{
download_data_server(id, data, include_codebook = TRUE)
}
\description{
Shiny module server for downloading data
}
