% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corp_design.R
\name{find_banner}
\alias{find_banner}
\title{FIND Banner}
\usage{
find_banner(title, extra_html = NULL)
}
\description{
FIND Banner
}
\examples{

page_title <- "Test Directory"
extra_html <- '<p>This fully searchable directory currently lists all 
                 commercially available immunoassays and molecular tests
               </p>'

ui <- find_dashboard_page(
 banner = find_banner(page_title, extra_html = extra_html),
 variants_ui("variants")
)

ui <- find_dashboard_page(
 title = "Test directory"
 variants_ui("variants")
)
}
