% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codebook.R
\name{get_codebook}
\alias{get_codebook}
\alias{get_codebook_}
\alias{get_codebook_extended}
\title{Get Codebook}
\usage{
get_codebook_(time = shinyfind:::today_at_sunrise())

get_codebook_extended(show_cacluation = FALSE)
}
\description{
Memoised Function, updates once a day
}
