% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\alias{get_data_}
\title{Universal Data Download}
\usage{
get_data_(
  dataset,
  repo = "FINDtestdirData",
  branch = "main",
  time = today_at_sunrise()
)
}
\description{
Memoised Function, updates once a day
}
