% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_over_group}
\alias{summarize_over_group}
\title{Step 2: Summarize Data Over Groups}
\usage{
summarize_over_group(data_summarized_over_time, group = NULL)
}
\description{
Step 2: Summarize Data Over Groups
}
