% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_testdir.R
\name{table_tests}
\alias{table_tests}
\title{Lists of Test Table for test directory apps}
\usage{
table_tests(session, data, meta_name, app = "covid19")
}
\description{
Lists of Test Table for test directory apps
}
\examples{

table_tests(session, data, meta_name)
}
