% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_over_time}
\alias{summarize_over_time}
\title{Summarize Data over Time and Groups}
\usage{
summarize_over_time(data_filtered)
}
\description{
All apps should use this functions to do aggregations!
}
\details{
In principle first it should be: summarise by time (examples time range given
below for reference) and then summarise by group.

The avg_cap_new_tests of the groups should be calculated using median and not
mean.

Problem is that if we calculate the means of groups using the sum of tests
divided by the sums of the population in all countries of the groups,
countries with low tests and high population skew downwards the mean and
vice versa.

See discussiofn in: https://github.com/dsbbfinddx/shinyfindapps/issues/142#issuecomment-1023435405
Step 1: Summarize Data Over Time
}
