% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_over_time}
\alias{summarize_over_time}
\alias{summarize_over_group}
\title{Summarize Data over Time and Groups}
\usage{
summarize_over_time(data_filtered)

summarize_over_group(data_summarized_over_time, group = NULL)
}
\description{
All apps should use this functions to do aggregations!
}
\details{
In principle first it should be: (1) summarise by time (examples time range given
below for reference) and then (2) summarise by group.

Problem is that if we calculate the means of groups using the sum of tests
divided by the sums of the population in all countries of the groups,
countries with low tests and high population skew downwards the mean and
vice versa.

See discussiofn in: https://github.com/finddx/shinyfindapps/issues/142#issuecomment-1023435405
}
\examples{

library(shinyfind)
library(dplyr)
. <- get_data_all()
data_all <- .$data_all
country_last_update_info <- .$country_last_update_info

time_start <- as.Date(max(data_all$time) - 367)
time_end <- as.Date(max(data_all$time) - 2)
data_filtered <-
  data_all |>
  filter(set == "country") |>
  select(-name, -set) |>
  left_join(country_last_update_info, by = "unit") |>
  filter(!is.na(last_update))|>
  filter(dplyr::between(
    time,
    as.Date(time_start),
    as.Date(time_end)
  ))
data_summarized_over_time <- shinyfind::summarize_over_time(data_filtered)
shinyfind::summarize_over_group(data_summarized_over_time, "income")
shinyfind::summarize_over_group(data_summarized_over_time, "who_region")
}
